cpg_rag_system/requirements.txt
# CPG RAG System - Python Dependencies
# Install: pip install -r requirements.txt --break-system-packages

# Core LangChain
langchain==0.1.0
langchain-community==0.0.13

# Vector Store
chromadb==0.4.22

# Graph Database
neo4j==5.16.0

# Embeddings & LLM (Ollama)
ollama==0.1.6

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
typing-extensions==4.9.0

# Data Processing
numpy==1.26.3

# Optional: Enhanced features
colorama==0.4.6  # Colored terminal output
rich==13.7.0     # Rich text formatting
tabulate==0.9.0  # Table formatting
